- Improved PhaseField Component Doxygen documentation
 - Fixed return type of amr_interpolator::coarse_value

 - Removed '_view' suffix from variable names in Benchmark applications of the
   PhaseField component

 - Renamed variable 'delta_u' to 'epsilon_u' and fixed error computations in
   Heat application of the PhaseField component

 - Fixed visibility of copy constructor of bc_fd< ScalarField<T>, STN, CC, F,
   BC::Dirichlet, 1 >

 - Fixed DSGN constexpr type in bc_fd for FCI conditions

 - Using scinew macro in PhaseField detail::bcfd_view and removed unneeded bci
   loop from constructor

 - Removed duplicate entry in PhaseField/BoundaryCondition/sub.mk file

 - Changed BCInterface from class to struct

 - Removed friendship from dw_view to virtual_view and made copy constructor
   protected so virtual_view can still access it

+++ src/CCA/Components/PhaseField/BoundaryConditions/detail/bc_fd_FineCoarseInterface_G1_FCBilinear.h
@@ -61,7 +61,7 @@ namespace detail
  */
 template<typename T, StnType STN, VarType VAR, Patch::FaceType F>
 class bc_fd < ScalarField<T>, STN, VAR, F, BC::FineCoarseInterface, 1, FC::FCBilinear >
-    : virtual public basic_fd_view< ScalarField<T>, STN >
+    : virtual public basic_fd_view < ScalarField<T>, STN >
 {
 private: // STATIC MEMBERS
 
@@ -78,7 +78,7 @@ private: // STATIC MEMBERS
     static constexpr int SGN = get_face<F>::sgn;
 
     /// Boundary face normal vector sign (double)
-    static constexpr int DSGN = get_face<F>::dsgn;
+    static constexpr double DSGN = get_face<F>::dsgn;
 
 private: // TYPES
 
@@ -427,6 +427,7 @@ public: // VIEW METHODS
                ( low[Y] <= id[Y] && id[Y] < high[Y] ) &&
                ( low[Z] <= id[Z] && id[Z] < high[Z] );
     };
+
 
 public: // BASIC FD VIEW METHODS
